feat: add useItems hook
