+++ src/components/MenuItem/index.tsx
@@ -1,19 +1,15 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { ICON_ARROW_RIGHT } from '~/constants/icons';
-// import { MenuListContext } from '~/menu/menu-context';
 import { useMenuItem } from '~/menu/use-menu-item';
-// import { MenuListContext } from '~/menu/menu-list-context';
-// import { useMenuItem } from '~/menu/use-menu-item';
-import { PopupXPosition } from '~/popup/popup-utils';
 import { mergeEvents, mergeRefs } from '~/utils/react';
 import { MenuListProps, MENU_PADDING_Y } from '../MenuList';
 import {
-  Accelerator,
   IconContainer,
   Label,
   StyledMenuItem,
   SubmenuIcon,
+  Accelerator,
 } from './style';
 
 export interface MenuItemProps extends React.HTMLAttributes<HTMLLIElement> {
@@ -40,10 +36,7 @@ export const MenuItem = React.forwardRef<HTMLLIElement, MenuItemProps>(
     },
     ref,
   ) => {
-    // const list = React.useContext(MenuListContext);
     const item = useMenuItem(!!submenu);
-
-    // const selected = list?.selectedItem?.ref.current === item.ref.current;
 
     const _submenu = React.useMemo(() => {
       if (!item.isSubmenuOpened || !submenu) return null;
